% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ErkerLinesOnMap.R
\docType{package}
\name{ErkerLinesOnMap}
\alias{ErkerLinesOnMap}
\alias{ErkerLinesOnMap-package}
\title{not sure}
\description{
Use the center of an SpatialLines object to obtain a google map of the area, plot the lines and map together
}
\examples{
#See Vignettes
}
\author{
Tedward Erker \email{erker@wisc.edu}
}

