% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Center.R
\name{get_map_center}
\alias{get_map_center}
\title{Get Center of  SpatialLines Object}
\usage{
get_map_center(x)
}
\arguments{
\item{x}{A ggplot}
}
\value{
Named Vector of center coordinates
}
\description{
find the middle of each axis
}

