% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PlotLinesOnMap.R
\name{plot_fortifiedLines_onGGMAP}
\alias{plot_fortifiedLines_onGGMAP}
\title{Plot SpatialLines Object on Google Map}
\usage{
plot_fortifiedLines_onGGMAP(fortified_spatialLines, zoom = 12, color = "bw",
  maptype = "terrain")
}
\arguments{
\item{fortified_spatialLines}{A spatialLines object that has been fortified by ggplot}

\item{zoom}{Input for ggmap's zoom}

\item{color}{Input for ggmap's color}

\item{maptype}{Inputr for ggmaps's maptype}
}
\value{
plot of map
}
\description{
plot fortified spatial lines object on google map
}
\examples{
# see vignette
 # plot_fortifiedLines_onGGMAP(street)
}

